Move CaptureRunnable to an independent file
